+++ frontend/src/Dashboard.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { getAccessKey, clearAuth, getUserInfo, parseJwt, type User as AuthUser } from './auth';
 import { useNavigate } from 'react-router-dom';
-import { LogOut, Key, Users, Copy, Trash2, Shield, Loader, Plus, User, Pencil, RotateCcw, ChevronLeft, ChevronRight, ArrowUpDown, ArrowUp, ArrowDown, Check, X as XIcon, BookOpen, Lock } from 'lucide-react';
+import { LogOut, Key, Users, Copy, Trash2, Shield, Loader, Plus, User, Pencil, RotateCcw, ChevronLeft, ChevronRight, ArrowUpDown, ArrowUp, ArrowDown, Check, X as XIcon, UserCog } from 'lucide-react';
 import axios from 'axios';
 
 
@@ -52,7 +52,6 @@ export const Dashboard = () => {
     const [users, setUsers] = useState<Record<string, UserDetails>>({});
     const [performanceMetrics, setPerformanceMetrics] = useState<Record<string, number>>({});
     const [performanceLabels, setPerformanceLabels] = useState<string[]>([]);
-    const [appInfo, setAppInfo] = useState<{ version: string, backend: string } | null>(null);
     const [loading, setLoading] = useState(true);
 
     // Key Modal State
@@ -111,11 +110,6 @@ export const Dashboard = () => {
         setUser(info);
         fetchData(info.is_admin);
 
-        // Fetch App Info
-        fetch('/api/app-info')
-            .then(res => res.json())
-            .then(data => setAppInfo(data))
-            .catch(err => console.error('Failed to fetch app info:', err));
 
         return () => {
             if (timer) clearTimeout(timer);
@@ -289,50 +283,6 @@ export const Dashboard = () => {
         setShowUserModal(true);
     };
 
-
-    // Security Settings State
-    const [passState, setPassState] = useState({ oldPass: '', newPass: '', confirmPass: '' });
-    const [emailState, setEmailState] = useState({ oldPass: '', newEmail: '', confirmEmail: '' });
-
-    const handleChangePassword = async (e: React.FormEvent) => {
-        e.preventDefault();
-        if (passState.newPass !== passState.confirmPass) {
-            alert("New passwords do not match!");
-            return;
-        }
-        try {
-            await axios.post('/api/change-password', {
-                oldPassword: passState.oldPass,
-                newPassword: passState.newPass
-            }, { headers: { Authorization: `Bearer ${getAccessKey()}` } });
-
-            alert("Password updated successfully.");
-            setPassState({ oldPass: '', newPass: '', confirmPass: '' });
-        } catch (e: any) {
-            const msg = e.response?.data ? String(e.response.data).trim() : 'Failed to update password';
-            alert(msg);
-        }
-    };
-
-    const handleChangeEmail = async (e: React.FormEvent) => {
-        e.preventDefault();
-        if (emailState.newEmail !== emailState.confirmEmail) {
-            alert("New email addresses do not match!");
-            return;
-        }
-        try {
-            await axios.post('/api/request-email-change', {
-                oldPassword: emailState.oldPass,
-                newEmail: emailState.newEmail
-            }, { headers: { Authorization: `Bearer ${getAccessKey()}` } });
-
-            alert("Confirmation email sent to the new address. Please check your inbox to finalize the change.");
-            setEmailState({ oldPass: '', newEmail: '', confirmEmail: '' });
-        } catch (e: any) {
-            const msg = e.response?.data ? String(e.response.data).trim() : 'Failed to request email change';
-            alert(msg);
-        }
-    };
     const handleSort = (type: 'keys' | 'users', key: string) => {
         let direction: 'asc' | 'desc' = 'asc';
         if (sortConfig.type === type && sortConfig.key === key && sortConfig.direction === 'asc') {
@@ -389,10 +339,16 @@ export const Dashboard = () => {
                         <span className="text-sm text-slate-400">{user.is_admin ? 'Administrator' : 'Standard User'}</span>
                     </div>
                 </div>
-                <button onClick={handleLogout} className="flex items-center gap-2 px-4 py-2 rounded-lg hover:bg-white/5 transition-colors text-slate-300">
-                    <LogOut size={18} />
-                    <span>Sign Out</span>
-                </button>
+                <div className="flex gap-3">
+                    <button onClick={() => navigate('/settings')} className="flex items-center gap-2 px-4 py-2 rounded-lg hover:bg-white/5 transition-colors text-slate-300">
+                        <UserCog size={18} />
+                        <span>Settings</span>
+                    </button>
+                    <button onClick={handleLogout} className="flex items-center gap-2 px-4 py-2 rounded-lg hover:bg-white/5 transition-colors text-slate-300">
+                        <LogOut size={18} />
+                        <span>Sign Out</span>
+                    </button>
+                </div>
             </div>
 
             {loading ? (
@@ -683,91 +639,6 @@ export const Dashboard = () => {
                 </div>
             )}
 
-            {/* Security Settings Panel */}
-            <div className="glass-panel p-6 col-span-1 lg:col-span-2">
-                <h2 className="text-xl font-semibold flex items-center gap-2 mb-6">
-                    <Lock className="text-indigo-400" /> Security Settings
-                </h2>
-                <div className="grid grid-cols-1 md:grid-cols-2 gap-8">
-                    {/* Change Password */}
-                    <div className="bg-white/5 rounded-lg p-6 border border-white/5">
-                        <h3 className="text-lg font-medium mb-4 text-slate-200">Change Password</h3>
-                        <form onSubmit={handleChangePassword} className="space-y-4">
-                            <div>
-                                <label className="block text-sm text-slate-400 mb-1">Current Password</label>
-                                <input
-                                    type="password" required
-                                    className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
-                                    value={passState.oldPass}
-                                    onChange={e => setPassState({ ...passState, oldPass: e.target.value })}
-                                />
-                            </div>
-                            <div>
-                                <label className="block text-sm text-slate-400 mb-1">New Password</label>
-                                <input
-                                    type="password" required minLength={8}
-                                    className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
-                                    value={passState.newPass}
-                                    onChange={e => setPassState({ ...passState, newPass: e.target.value })}
-                                />
-                            </div>
-                            <div>
-                                <label className="block text-sm text-slate-400 mb-1">Confirm New Password</label>
-                                <input
-                                    type="password" required minLength={8}
-                                    className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
-                                    value={passState.confirmPass}
-                                    onChange={e => setPassState({ ...passState, confirmPass: e.target.value })}
-                                />
-                            </div>
-                            <div className="pt-2">
-                                <button type="submit" className="w-full bg-indigo-600 hover:bg-indigo-500 text-white px-4 py-2 rounded-lg transition-colors">
-                                    Update Password
-                                </button>
-                            </div>
-                        </form>
-                    </div>
-
-                    {/* Change Email */}
-                    <div className="bg-white/5 rounded-lg p-6 border border-white/5">
-                        <h3 className="text-lg font-medium mb-4 text-slate-200">Change Registered Email</h3>
-                        <form onSubmit={handleChangeEmail} className="space-y-4">
-                            <div>
-                                <label className="block text-sm text-slate-400 mb-1">Current Password</label>
-                                <input
-                                    type="password" required
-                                    className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
-                                    value={emailState.oldPass}
-                                    onChange={e => setEmailState({ ...emailState, oldPass: e.target.value })}
-                                />
-                            </div>
-                            <div>
-                                <label className="block text-sm text-slate-400 mb-1">New Email Address</label>
-                                <input
-                                    type="email" required
-                                    className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
-                                    value={emailState.newEmail}
-                                    onChange={e => setEmailState({ ...emailState, newEmail: e.target.value })}
-                                />
-                            </div>
-                            <div>
-                                <label className="block text-sm text-slate-400 mb-1">Confirm New Email</label>
-                                <input
-                                    type="email" required
-                                    className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
-                                    value={emailState.confirmEmail}
-                                    onChange={e => setEmailState({ ...emailState, confirmEmail: e.target.value })}
-                                />
-                            </div>
-                            <div className="pt-2">
-                                <button type="submit" className="w-full bg-indigo-600 hover:bg-indigo-500 text-white px-4 py-2 rounded-lg transition-colors">
-                                    Request Email Change
-                                </button>
-                            </div>
-                        </form>
-                    </div>
-                </div>
-            </div>
             {user.is_admin && (
                 <div className="glass-panel p-6 col-span-1 lg:col-span-2">
                     <div className="flex justify-between items-center mb-6">
@@ -809,216 +680,186 @@ export const Dashboard = () => {
                 </div>
             )}
 
-            {/* API Info Panel */}
-            {appInfo && (
-                <div className="glass-panel p-6 col-span-1 lg:col-span-2">
-                    <h2 className="text-xl font-semibold flex items-center gap-2 mb-6">
-                        <BookOpen className="text-indigo-400" /> API Documentation
-                    </h2>
-                    <div className="flex flex-col items-center justify-center gap-6 py-4">
-                        <div className="text-center">
-                            <p className="text-slate-400 mb-2">Swagger Interface</p>
-                            <a
-                                href={`${appInfo.backend}/swagger/`}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                className="text-xl text-indigo-400 hover:text-indigo-300 underline decoration-indigo-500/50 underline-offset-4 font-medium transition-colors"
-                            >
-                                {appInfo.backend}/swagger/
-                            </a>
-                        </div>
-                        <div className="text-center w-full border-t border-white/5 pt-6">
-                            <p style={{ fontSize: '0.8rem' }} className="text-slate-500">
-                                Build {appInfo.version} | <a href="https://github.com/iulianpascalau/mx-epoch-proxy-go" className="hover:text-slate-400 underline decoration-slate-600 underline-offset-2" target="_blank" rel="noopener noreferrer">Solution</a>
-                            </p>
-                        </div>
 
-                    </div>
-                </div>
-            )}
-        </div>
-    )
-}
-
-{/* Key Modal */ }
-{
-    showKeyModal && (
-        <div className="fixed inset-0 bg-black/80 backdrop-blur-sm flex items-center justify-center p-4 z-50">
-            <div className="glass-panel w-full max-w-md p-6 animate-in fade-in zoom-in duration-200">
-                <h3 className="text-xl font-bold mb-4">Generate New Key</h3>
-                <form onSubmit={handleCreateKey}>
-                    <div className="mb-4">
-                        <label className="block text-sm text-slate-400 mb-1">Key Value (Optional)</label>
-                        <input
-                            type="text"
-                            className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
-                            placeholder="Leave empty for random UUID"
-                            value={newKeyVal}
-                            onChange={e => setNewKeyVal(e.target.value)}
-                        />
-                    </div>
-                    <div className="flex justify-end gap-3 mt-6">
-                        <button type="button" onClick={() => { setShowKeyModal(false); setNewKeyVal(''); }} className="px-4 py-2 hover:bg-white/5 rounded text-slate-300">Cancel</button>
-                        <button type="submit" className="px-4 py-2 bg-indigo-600 hover:bg-indigo-500 rounded text-white">Create</button>
-                    </div>
-                </form>
-            </div>
-        </div>
-    )
-}
-
-{/* User Modal */ }
-{
-    showUserModal && (
-        <div className="fixed inset-0 bg-black/80 backdrop-blur-sm flex items-center justify-center p-4 z-50">
-            <div className="glass-panel w-full max-w-md p-6 animate-in fade-in zoom-in duration-200">
-                <h3 className="text-xl font-bold mb-4">{isEditingUser ? 'Edit User' : 'Add New User'}</h3>
-                <form onSubmit={handleCreateUser}>
-                    <div className="space-y-4">
-                        <div>
-                            <label className="block text-sm text-slate-400 mb-1">Username</label>
-                            <input
-                                type="text" required
-                                readOnly={isEditingUser}
-                                className={`w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none ${isEditingUser ? 'opacity-50 cursor-not-allowed' : ''}`}
-                                value={newUserState.username}
-                                onChange={e => setNewUserState({ ...newUserState, username: e.target.value })}
-                            />
-                        </div>
-                        <div>
-                            <label className="block text-sm text-slate-400 mb-1">Password {isEditingUser && '(Leave empty to keep current)'}</label>
-                            <input
-                                type="password"
-                                required={!isEditingUser}
-                                className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none"
-                                value={newUserState.password}
-                                onChange={e => setNewUserState({ ...newUserState, password: e.target.value })}
-                            />
-                        </div>
-                        <div>
-                            <label className="block text-sm text-slate-400 mb-1">Max Requests (0 = Unlimited)</label>
-                            <input
-                                type="number" min="0"
-                                className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none"
-                                value={newUserState.maxRequests}
-                                onChange={e => setNewUserState({ ...newUserState, maxRequests: parseInt(e.target.value) })}
-                            />
-                        </div>
-                        <div>
-                            <label className="block text-sm text-slate-400 mb-1">Account Type</label>
-                            <select
-                                className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none"
-                                value={newUserState.accountType}
-                                onChange={e => setNewUserState({ ...newUserState, accountType: e.target.value })}
-                            >
-                                <option value="free">Free</option>
-                                <option value="premium">Premium</option>
-                            </select>
-                        </div>
-                        <div className="flex items-center gap-2">
-                            <input
-                                type="checkbox" id="isAdminCheck"
-                                className="w-4 h-4 rounded border-slate-700 bg-slate-800 text-emerald-600 focus:ring-emerald-500"
-                                checked={newUserState.isAdmin}
-                                onChange={e => setNewUserState({ ...newUserState, isAdmin: e.target.checked })}
-                            />
-                            <label htmlFor="isAdminCheck" className="text-sm text-slate-300">Grant Administrator Privileges</label>
+            {/* Key Modal */}
+            {
+                showKeyModal && (
+                    <div className="fixed inset-0 bg-black/80 backdrop-blur-sm flex items-center justify-center p-4 z-50">
+                        <div className="glass-panel w-full max-w-md p-6 animate-in fade-in zoom-in duration-200">
+                            <h3 className="text-xl font-bold mb-4">Generate New Key</h3>
+                            <form onSubmit={handleCreateKey}>
+                                <div className="mb-4">
+                                    <label className="block text-sm text-slate-400 mb-1">Key Value (Optional)</label>
+                                    <input
+                                        type="text"
+                                        className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
+                                        placeholder="Leave empty for random UUID"
+                                        value={newKeyVal}
+                                        onChange={e => setNewKeyVal(e.target.value)}
+                                    />
+                                </div>
+                                <div className="flex justify-end gap-3 mt-6">
+                                    <button type="button" onClick={() => { setShowKeyModal(false); setNewKeyVal(''); }} className="px-4 py-2 hover:bg-white/5 rounded text-slate-300">Cancel</button>
+                                    <button type="submit" className="px-4 py-2 bg-indigo-600 hover:bg-indigo-500 rounded text-white">Create</button>
+                                </div>
+                            </form>
                         </div>
                     </div>
-                    <div className="flex justify-end gap-3 mt-6">
-                        <button type="button" onClick={() => setShowUserModal(false)} className="px-4 py-2 hover:bg-white/5 rounded text-slate-300">Cancel</button>
-                        <button type="submit" className="px-4 py-2 bg-emerald-600 hover:bg-emerald-500 rounded text-white">{isEditingUser ? 'Update User' : 'Save User'}</button>
+                )
+            }
+
+            {/* User Modal */}
+            {
+                showUserModal && (
+                    <div className="fixed inset-0 bg-black/80 backdrop-blur-sm flex items-center justify-center p-4 z-50">
+                        <div className="glass-panel w-full max-w-md p-6 animate-in fade-in zoom-in duration-200">
+                            <h3 className="text-xl font-bold mb-4">{isEditingUser ? 'Edit User' : 'Add New User'}</h3>
+                            <form onSubmit={handleCreateUser}>
+                                <div className="space-y-4">
+                                    <div>
+                                        <label className="block text-sm text-slate-400 mb-1">Username</label>
+                                        <input
+                                            type="text" required
+                                            readOnly={isEditingUser}
+                                            className={`w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none ${isEditingUser ? 'opacity-50 cursor-not-allowed' : ''}`}
+                                            value={newUserState.username}
+                                            onChange={e => setNewUserState({ ...newUserState, username: e.target.value })}
+                                        />
+                                    </div>
+                                    <div>
+                                        <label className="block text-sm text-slate-400 mb-1">Password {isEditingUser && '(Leave empty to keep current)'}</label>
+                                        <input
+                                            type="password"
+                                            required={!isEditingUser}
+                                            className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none"
+                                            value={newUserState.password}
+                                            onChange={e => setNewUserState({ ...newUserState, password: e.target.value })}
+                                        />
+                                    </div>
+                                    <div>
+                                        <label className="block text-sm text-slate-400 mb-1">Max Requests (0 = Unlimited)</label>
+                                        <input
+                                            type="number" min="0"
+                                            className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none"
+                                            value={newUserState.maxRequests}
+                                            onChange={e => setNewUserState({ ...newUserState, maxRequests: parseInt(e.target.value) })}
+                                        />
+                                    </div>
+                                    <div>
+                                        <label className="block text-sm text-slate-400 mb-1">Account Type</label>
+                                        <select
+                                            className="w-full bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-emerald-500 focus:outline-none"
+                                            value={newUserState.accountType}
+                                            onChange={e => setNewUserState({ ...newUserState, accountType: e.target.value })}
+                                        >
+                                            <option value="free">Free</option>
+                                            <option value="premium">Premium</option>
+                                        </select>
+                                    </div>
+                                    <div className="flex items-center gap-2">
+                                        <input
+                                            type="checkbox" id="isAdminCheck"
+                                            className="w-4 h-4 rounded border-slate-700 bg-slate-800 text-emerald-600 focus:ring-emerald-500"
+                                            checked={newUserState.isAdmin}
+                                            onChange={e => setNewUserState({ ...newUserState, isAdmin: e.target.checked })}
+                                        />
+                                        <label htmlFor="isAdminCheck" className="text-sm text-slate-300">Grant Administrator Privileges</label>
+                                    </div>
+                                </div>
+                                <div className="flex justify-end gap-3 mt-6">
+                                    <button type="button" onClick={() => setShowUserModal(false)} className="px-4 py-2 hover:bg-white/5 rounded text-slate-300">Cancel</button>
+                                    <button type="submit" className="px-4 py-2 bg-emerald-600 hover:bg-emerald-500 rounded text-white">{isEditingUser ? 'Update User' : 'Save User'}</button>
+                                </div>
+                            </form>
+                        </div>
                     </div>
-                </form>
-            </div>
-        </div>
-    )
-}
+                )
+            }
 
-{/* Manage User Keys Modal */ }
-{
-    managingKeysForUser && (
-        <div className="fixed inset-0 bg-black/80 backdrop-blur-sm flex items-center justify-center p-4 z-50">
-            <div className="glass-panel w-full max-w-2xl p-6 animate-in fade-in zoom-in duration-200">
-                <div className="flex justify-between items-center mb-6">
-                    <h3 className="text-xl font-bold flex items-center gap-2">
-                        <Key className="text-indigo-400" />
-                        Keys for {managingKeysForUser.Username}
-                    </h3>
-                    <button
-                        onClick={() => setManagingKeysForUser(null)}
-                        className="text-slate-400 hover:text-white"
-                    >
-                        ✕
-                    </button>
-                </div>
+            {/* Manage User Keys Modal */}
+            {
+                managingKeysForUser && (
+                    <div className="fixed inset-0 bg-black/80 backdrop-blur-sm flex items-center justify-center p-4 z-50">
+                        <div className="glass-panel w-full max-w-2xl p-6 animate-in fade-in zoom-in duration-200">
+                            <div className="flex justify-between items-center mb-6">
+                                <h3 className="text-xl font-bold flex items-center gap-2">
+                                    <Key className="text-indigo-400" />
+                                    Keys for {managingKeysForUser.Username}
+                                </h3>
+                                <button
+                                    onClick={() => setManagingKeysForUser(null)}
+                                    className="text-slate-400 hover:text-white"
+                                >
+                                    ✕
+                                </button>
+                            </div>
 
-                {/* Add Key Form */}
-                <form onSubmit={handleCreateKey} className="mb-6 flex gap-2">
-                    <input
-                        type="text"
-                        className="flex-1 bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
-                        placeholder="New Key Value (Optional)"
-                        value={newKeyVal}
-                        onChange={e => setNewKeyVal(e.target.value)}
-                    />
-                    <button type="submit" className="px-4 py-2 bg-indigo-600 hover:bg-indigo-500 rounded text-white whitespace-nowrap">
-                        Add Key
-                    </button>
-                </form>
+                            {/* Add Key Form */}
+                            <form onSubmit={handleCreateKey} className="mb-6 flex gap-2">
+                                <input
+                                    type="text"
+                                    className="flex-1 bg-slate-800 border border-slate-700 rounded p-2 text-slate-200 focus:ring-2 focus:ring-indigo-500 focus:outline-none"
+                                    placeholder="New Key Value (Optional)"
+                                    value={newKeyVal}
+                                    onChange={e => setNewKeyVal(e.target.value)}
+                                />
+                                <button type="submit" className="px-4 py-2 bg-indigo-600 hover:bg-indigo-500 rounded text-white whitespace-nowrap">
+                                    Add Key
+                                </button>
+                            </form>
 
-                {/* Keys List */}
-                <div className="overflow-x-auto max-h-[400px] overflow-y-auto">
-                    <table className="w-full text-left border-collapse">
-                        <thead>
-                            <tr className="border-b border-white/10 text-slate-400 text-sm uppercase">
-                                <th className="py-2 px-4">Key Value</th>
-                                <th className="py-2 px-4">Requests</th>
-                                <th className="py-2 px-4 text-right">Actions</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {Object.entries(keys)
-                                .filter(([_, details]) => details.Username === managingKeysForUser.Username)
-                                .map(([k, details]) => (
-                                    <tr key={k} className="border-b border-white/5 hover:bg-white/5 transition-colors">
-                                        <td className="py-2 px-4 font-mono text-sm text-indigo-200">
-                                            <div className="flex items-center gap-2">
-                                                {k}
-                                                <Copy
-                                                    size={12}
-                                                    className="cursor-pointer text-slate-500 hover:text-white"
-                                                    onClick={() => copyToClipboard(k)}
-                                                />
-                                            </div>
-                                        </td>
-                                        <td className="py-2 px-4 text-slate-300">
-                                            {details.KeyCounter}
-                                        </td>
-                                        <td className="py-2 px-4 text-right">
-                                            <button
-                                                onClick={() => handleDeleteKey(k, managingKeysForUser.Username)}
-                                                className="text-red-400 hover:text-red-300 p-1 rounded hover:bg-red-400/10 transition-colors"
-                                            >
-                                                <Trash2 size={14} />
-                                            </button>
-                                        </td>
-                                    </tr>
-                                ))}
-                            {Object.values(keys).filter(k => k.Username === managingKeysForUser.Username).length === 0 && (
-                                <tr>
-                                    <td colSpan={3} className="py-4 text-center text-slate-500">
-                                        No specific keys for this user.
-                                    </td>
-                                </tr>
-                            )}
-                        </tbody>
-                    </table>
-                </div>
-            </div>
-        </div>
-    )
-}
+                            {/* Keys List */}
+                            <div className="overflow-x-auto max-h-[400px] overflow-y-auto">
+                                <table className="w-full text-left border-collapse">
+                                    <thead>
+                                        <tr className="border-b border-white/10 text-slate-400 text-sm uppercase">
+                                            <th className="py-2 px-4">Key Value</th>
+                                            <th className="py-2 px-4">Requests</th>
+                                            <th className="py-2 px-4 text-right">Actions</th>
+                                        </tr>
+                                    </thead>
+                                    <tbody>
+                                        {Object.entries(keys)
+                                            .filter(([_, details]) => details.Username === managingKeysForUser.Username)
+                                            .map(([k, details]) => (
+                                                <tr key={k} className="border-b border-white/5 hover:bg-white/5 transition-colors">
+                                                    <td className="py-2 px-4 font-mono text-sm text-indigo-200">
+                                                        <div className="flex items-center gap-2">
+                                                            {k}
+                                                            <Copy
+                                                                size={12}
+                                                                className="cursor-pointer text-slate-500 hover:text-white"
+                                                                onClick={() => copyToClipboard(k)}
+                                                            />
+                                                        </div>
+                                                    </td>
+                                                    <td className="py-2 px-4 text-slate-300">
+                                                        {details.KeyCounter}
+                                                    </td>
+                                                    <td className="py-2 px-4 text-right">
+                                                        <button
+                                                            onClick={() => handleDeleteKey(k, managingKeysForUser.Username)}
+                                                            className="text-red-400 hover:text-red-300 p-1 rounded hover:bg-red-400/10 transition-colors"
+                                                        >
+                                                            <Trash2 size={14} />
+                                                        </button>
+                                                    </td>
+                                                </tr>
+                                            ))}
+                                        {Object.values(keys).filter(k => k.Username === managingKeysForUser.Username).length === 0 && (
+                                            <tr>
+                                                <td colSpan={3} className="py-4 text-center text-slate-500">
+                                                    No specific keys for this user.
+                                                </td>
+                                            </tr>
+                                        )}
+                                    </tbody>
+                                </table>
+                            </div>
+                        </div>
+                    </div>
+                )
+            }
         </div >
     );
 };
